+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Complete parsing logic for FQNs to match with the Rust spec.
 - Rename the `Fqn` type to `SimplePath` to have the same naming as the Rust spec.
+- Expose all fields of the `Index` struct as the data can be changed through `serde` anyways.
 
 ## [0.1.1] - 2021-07-30