+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+## [0.2.0] - 2021-07-30
+
 ### Added
 
 - Paths that only contain the crate name with no item (like just `anyhow` or `std`), will now
@@ -33,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/dnaka91/docsearch/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/dnaka91/docsearch/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/dnaka91/docsearch/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/dnaka91/docsearch/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/dnaka91/docsearch/releases/tag/v0.1.0